+++ src/main/java/no/nav/pto/veilarbportefolje/aap/AapClient.kt
@@ -1,25 +1,23 @@
-package no.nav.pto.veilarbportefolje.aap.controller
+package no.nav.pto.veilarbportefolje.aap
 
 import no.nav.common.rest.client.RestClient.baseClient
 import no.nav.common.rest.client.RestUtils
-import no.nav.common.rest.client.RestUtils.MEDIA_TYPE_JSON
 import no.nav.common.utils.UrlUtils
-import no.nav.pto.veilarbportefolje.aap.client.AapIPeriodeRequest
-import no.nav.pto.veilarbportefolje.aap.client.AapRequest
-import no.nav.pto.veilarbportefolje.aap.client.AapResponseDto
-import no.nav.pto.veilarbportefolje.aap.client.AapResponseMaksimumDto
+import no.nav.pto.veilarbportefolje.aap.domene.AapSakRequest
+import no.nav.pto.veilarbportefolje.aap.domene.AapSakResponseDto
+import no.nav.pto.veilarbportefolje.aap.domene.AapVedtakRequest
+import no.nav.pto.veilarbportefolje.aap.domene.AapVedtakResponseDto
 import no.nav.pto.veilarbportefolje.util.deserializeJsonOrThrow
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import org.springframework.http.HttpHeaders
-import java.time.LocalDate
 import java.util.function.Supplier
 
 class AapClient(private val baseUrl: String, private val machineToMachineTokenSupplier: Supplier<String>) {
     private val client: OkHttpClient = baseClient()
 
-    fun hentAapForPersonnr(personnr: String): List<AapResponseDto> {
-        val requestBody = AapRequest(listOf(personnr))
+    fun hentAapSak(personnr: String): List<AapSakResponseDto> {
+        val requestBody = AapSakRequest(listOf(personnr))
 
         val request = Request.Builder()
             .url(UrlUtils.joinPaths(baseUrl, "/kelvin/sakerByFnr"))
@@ -35,8 +33,8 @@ class AapClient(private val baseUrl: String, private val machineToMachineTokenSu
     }
 
 
-    fun hentAapMaximum(personnr: String, fom: String, tom: String): AapResponseMaksimumDto {
-        val requestBody = AapIPeriodeRequest(personnr, fom, tom)
+    fun hentAapVedtak(personnr: String, fom: String, tom: String): AapVedtakResponseDto {
+        val requestBody = AapVedtakRequest(personnr, fom, tom)
 
         val request = Request.Builder()
             .url(UrlUtils.joinPaths(baseUrl, "/kelvin/maksimumUtenUtbetaling"))
@@ -52,3 +50,4 @@ class AapClient(private val baseUrl: String, private val machineToMachineTokenSu
 
     }
 }
+
